PT Card List with Card Sets
A list of PT Cards in Perfect Team 23 with additional columns to show if a player is packable, a mission reward, tournament reward (and which TE set), platinum series card or limited edition. Designed to allow Theme Team Crew (TTC) members to determine how easily obtaining certain cards are for their teams.